# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Diliké",
        "version": "1.0"
    },
    "paths": {
...
        "/api/restaurants": {
            "get": {
                "tags": [
                    "restaurants"
                ],
                "summary": "liste des restaurants",
                "description": "liste de de tous les restaurants",
                "operationId": "restaurants_list",
                "responses": {
                    "200": {
                        "description": "Afficher la liste de de tous les restaurants"
                    }
                }
            },
            "post": {
                "tags": [
                    "restaurants"
                ],
                "summary": "Créer un restaurant",
                "description": "Inscription d'un restaurant",
                "operationId": "restaurants",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string"
                                    },
                                    "adresse": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "nom": "Les Delices",
                                    "adresse": "Yopougon-Maroc",
                                    "email": "delice@gmail.com",
                                    "phone": "0102030405"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/restaurants/{id}": {
            "get": {
                "tags": [
                    "restaurants"
                ],
                "summary": "Obtenir un seul restaurant",
                "description": "Obetenir un seul restaurant",
                "operationId": "one_restaurant",
                "responses": {
                    "200": {
                        "description": "Afficher les information d'un seul restaurant"
                    }
                }
            },
            "put": {
                "tags": [
                    "restaurants"
                ],
                "summary": "Mettre à jour un restaurant",
                "description": "Mise à jour d'un restaurant",
                "operationId": "_update_restaurants",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string"
                                    },
                                    "adresse": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "nom": "Les Delices",
                                    "adresse": "Yopougon-Maroc",
                                    "email": "delice@gmail.com",
                                    "phone": "0102030405"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "delete": {
                "tags": [
                    "restaurants"
                ],
                "summary": "Supprimer un restaurant",
                "description": "Suprrime un restaurant",
                "operationId": "delete_restaurants",
                "responses": {
                    "200": {
                        "description": "Supprime une restaurant"
                    }
                }
            }
        },
...
    }
}